requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
scikit-image>=0.22.0
wandb>=0.17.0
PyYAML>=6.0.0
